requirements.txt
Django
python-dotenv
psycopg[binary]
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-tenants